Add support for attaching content in clipboard
This adds a feature that when in normal mode, pressing `p` or `CTRL+V` will paste the current
text from clipboard, or if it's file (such as an image) it will attach it to the message.

A limitation with this solution is that attached files will be stored in a temporary file,
so they are not persisant. I'm open for suggestions on a better way to store them.

Also, I wasn't able to bind CTRL+Shift+V in normal mode, but I think this paste method
should be used there as well.